requirements.txt
flask==2.3.2
pandas==2.0.3
psycopg2-binary == 2.9.6
sqlalchemy==2.0.19
